backend/services/cust_context/pyproject.toml
[tool.poetry]
name = "template-service-python-prisma"
version = "0.1.0"
description = "MilkyHoop Template Service Python Prisma"
authors = ["MilkyHoop Team"]

[tool.poetry.dependencies]
python = "^3.10"
redis = "^4.5.4"
aiohttp = "^3.8.0"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
grpcio = "^1.71.0"
grpcio-tools = "^1.63.0"
grpcio-health-checking = "^1.63.0"
protobuf = "^5.29.0"
grpc-interceptor = "^0.15.0"
structlog = "^25.2.0"
tenacity = "^8.2.3"
python-dotenv = "^1.0.1"
prometheus-client = "^0.20.0"
prisma = "0.15.0"
numpy = "^1.26.0"
grpcio-status = "1.71.0"

[tool.poetry.dev-dependencies]
mypy = "^1.9.0"
pytest = "^8.1.1"
ruff = "^0.3.4"
pytest-asyncio = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

